requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.3
google-genai==1.14.0
Pillow==10.1.0
python-dateutil==2.8.2 